[change] adapt to the latest version of chainer
Chainer 6.7.0 requires typing-sextensions 3.6.0, which is incompatble with torch 1.9.0 at runtime (in fact, the requirements of PyTorch's is loose). On the other hand, chainer 7.8.0 is compatible with typing-extensions 3.10.0.0 and hence with the latest torch.

Some package importing needs to be updated to catch up with the latest version of chainer. As far as the older versions of docs suggest, the changes this commit makes should cause no problems and carry over to v.2.0.2.